+++ src/account/MintTransaction.hpp
@@ -17,12 +17,15 @@ namespace sgns
     class MintTransaction : public IGeniusTransactions
     {
     public:
-        MintTransaction( double new_amount, std::string chainid, std::string tokenid,  const SGTransaction::DAGStruct &dag );
+        MintTransaction( uint64_t                        new_amount,
+                         std::string                     chainid,
+                         std::string                     tokenid,
+                         const SGTransaction::DAGStruct &dag );
         ~MintTransaction() override = default;
 
         std::vector<uint8_t>                    SerializeByteVector() override;
         static std::shared_ptr<MintTransaction> DeSerializeByteVector( const std::vector<uint8_t> &data );
-        double                                GetAmount() const;
+        uint64_t                                GetAmount() const;
 
         std::string GetTransactionSpecificPath() override
         {
@@ -30,7 +33,7 @@ namespace sgns
         }
 
     private:
-        double amount;
+        uint64_t    amount;
         std::string chain_id;
         std::string token_id;
